+++ autoarray/dataset/mock/mock_dataset.py
@@ -1,0 +1,6 @@
+class MockDataset:
+    def __init__(self, grid_inversion=None, psf=None, mask=None):
+
+        self.grid_inversion = grid_inversion
+        self.psf = psf
+        self.mask = mask
