+++ autoarray/mask/mock/mock_mask.py
@@ -1,0 +1,4 @@
+class MockMask:
+    def __init__(self, native_index_for_slim_index=None):
+
+        self.native_index_for_slim_index = native_index_for_slim_index
