+++ test_autoarray/fit/test_fit_imaging.py
@@ -1,8 +1,9 @@
-import autoarray as aa
-
 import numpy as np
 
-from autoarray.mock.mock import MockInversion, MockFitImaging
+import autoarray as aa
+
+from autoarray.inversion.inversion.mock.mock_inversion import MockInversion
+from autoarray.fit.mock.mock_fit_imaging import MockFitImaging
 
 
 def test__image_and_model_are_identical__no_masking__check_values_are_correct():
