+++ test_autoarray/inversion/inversion/test_matrices.py
@@ -4,7 +4,7 @@ import pytest
 
 import autoarray as aa
 
-from autoarray.mock.mock import MockMapper, MockLEq, MockInversion
+from autoarray.inversion.inversion.mock.mock_inversion import MockInversion
 
 
 directory = path.dirname(path.realpath(__file__))
